# RFC: Persist more objs as pure json