@adrinjalali 2023-03-30
Looking at it, I think I see a bug:

```
                         'min_impurity_decrease': {'__class__': 'str',
                                                   '__id__': 140450862391504,
                                                   '__loader__': 'JsonNode',
                                                   '__module__': 'builtins',
                                                   'content': '0.0',
                                                   'is_json': True},
```
why is the type `str` here? It should be `float` I think, shouldn't it?

What I'm thinking is, here with the change we lose dtypes of the objects, but I see now the dtypes are wrong as they are. Hmm...