@adrinjalali 2023-03-31
What I mean is, ideally we'd let json read the object, but then cast it with the type given in the type slot. Otherwise we're probably changing the types of numbers.