@BenjaminBossan 2023-03-31
> What I mean is, ideally we'd let json read the object, but then cast it with the type given in the type slot. Otherwise we're probably changing the types of numbers.

I see. Yes, I was wondering if may sometimes get the wrong numerical type, but at least our unit tests, where we check type, don't detect any incorrect conversion. If we do want to coerce types, we can't proceed with the refactor, no matter which version.