@adrinjalali 2023-03-31
We can probably at least optimize the dict node, so that we dump the values with json, but we keep dtypes of values as well as keys maybe?