@BenjaminBossan 2023-03-31
We could do that for `DictNode` (and also for `ListNode`, which may also persist valid json). If I understand your suggestion correctly, though, the end result would not be much more compact or readable than the existing solution. So maybe let's just stick with the existing solution? And maybe set the actual value's type in `__class__` instead of `str` (which would be BC breaking, so require to preserve the old implementation).