@adrinjalali 2023-03-31
Yeah I think that'd make sense.